+++ CHANGELOG.md
@@ -9,21 +9,42 @@ adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [unreleased]
 
+## [0.0.3] - 2025-06-08
+
+### Added
+
+- `media.archive_dashcam_footage`
+  - Added `container` parameter (defaults to `matroska`). Must match extension.
+  - Added `extension` parameter (defaults to `'mkv'`).
+  - Added `keep_audio` parameter (defaults to `False`).
+
+### Changed
+
+- `encode-dashcam`
+  - Set default `--video-max-bitrate` to `'30M'`.
+  - Set default encoder to `'libx265'`.
+- `media.archive_dashcam_footage`
+  - Set default `video_encoder` to `'libx265'`.
+
 ## [0.0.2] - 2025-06-07
+
+### Added
+
+- `encode-dashcam`
+  - Added `-crf` option.
+  - Added `--no-delete` option.
+- `media.archive_dashcam_footage`
+  - Added `crf` parameter for software encoders.
 
 ### Fixed
 
-- `media.archive_dashcam_footage`: was completely broken due to path issue.
+- `media.archive_dashcam_footage` was completely broken due to path issue.
 
 ### Changed
 
 - `encode-dashcam`
   - Set default level to `'auto'` (NVENC HEVC).
-  - Added `-crf` option.
-  - Added `--no-delete` option.
 - `media.archive_dashcam_footage`
-  - Added `crf` parameter for software encoders.
-  - Set default level to `'auto'` (NVENC HEVC)
   - Now accepts `Pattern[str]` for the `match_re` parameter.
   - Improved handling of encoder-specific arguments.
   - Added `no_delete` parameter.
@@ -32,4 +53,4 @@ adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 First version. `check_bookmarks_html_urls` may have unresolved issues.
 
-[unreleased]: https://github.com/Tatsh/deltona/compare/v0.0.2...HEAD
+[unreleased]: https://github.com/Tatsh/deltona/compare/v0.0.3...HEAD